pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aas-standard-parser"
version = "0.2.5"
description = "Some auxiliary functions for parsing standard submodels"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Daniel Klein", email = "daniel.klein@em.ag" }]
dependencies = ["typing>=3.7.4.3", "basyx-python-sdk>=1.2.1"]
[project.urls]
Homepage = "https://github.com/fluid40/aas-http-client"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"

[tool.commitizen.branch]
allowed = "^(main|master|develop|feature/.*|fix/.*|hotfix/.*)$"
